Split datastore from plugin into own module
